Use futurize to have python 3 compatible int instead of long

+++ src/plugin/plugin.py
@@ -493,7 +493,7 @@ class KodiVideoPlayer(InfoBarBase, InfoBarShowHide, SubsSupportStatus, SubsSuppo
             self.__timer.start(500, True)
         else:
             Notifications.RemovePopup(self.RESUME_POPUP_ID)
-            self.doSeek(long(self.__position))
+            self.doSeek(int(self.__position))
 
     def setImage(self, image):
         self.__image = image
